requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
common_tool==0.0.3
cryptography==3.4.7
Django==5.0.6
django-redis==5.4.0
idna==3.7
pycparser==2.22
pycryptodome==3.9.0
PyMySQL==1.1.1
redis==5.0.7
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.2
